# microservices/mano-dns/mano-dns-pdns-aggregator/mano-pdns-api/src/main/java/com/ubiqube/mano/pdns/model/ConfigSetting.java
/*
 * PowerDNS Authoritative HTTP API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.15
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package com.ubiqube.mano.pdns.model;

import java.util.Objects;

import org.jspecify.annotations.Nullable;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;

/**
 * ConfigSetting
 */
@JsonPropertyOrder({
		ConfigSetting.JSON_PROPERTY_NAME,
		ConfigSetting.JSON_PROPERTY_TYPE,
		ConfigSetting.JSON_PROPERTY_VALUE
})
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2025-05-23T10:34:53.990305454+02:00[Europe/Paris]", comments = "Generator version: 7.13.0")
public class ConfigSetting {
	public static final String JSON_PROPERTY_NAME = "name";
	@Nullable
	private String name;

	public static final String JSON_PROPERTY_TYPE = "type";
	@Nullable
	private String type;

	public static final String JSON_PROPERTY_VALUE = "value";
	@Nullable
	private String value;

	public ConfigSetting() {
	}

	public ConfigSetting name(@Nullable final String name) {

		this.name = name;
		return this;
	}

	/**
	 * set to \&quot;ConfigSetting\&quot;
	 *
	 * @return name
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_NAME)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getName() {
		return name;
	}

	@JsonProperty(JSON_PROPERTY_NAME)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setName(@Nullable final String name) {
		this.name = name;
	}

	public ConfigSetting type(@Nullable final String type) {

		this.type = type;
		return this;
	}

	/**
	 * The name of this setting (e.g. ‘webserver-port’)
	 *
	 * @return type
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getType() {
		return type;
	}

	@JsonProperty(JSON_PROPERTY_TYPE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setType(@Nullable final String type) {
		this.type = type;
	}

	public ConfigSetting value(@Nullable final String value) {

		this.value = value;
		return this;
	}

	/**
	 * The value of setting name
	 *
	 * @return value
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_VALUE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getValue() {
		return value;
	}

	@JsonProperty(JSON_PROPERTY_VALUE)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setValue(@Nullable final String value) {
		this.value = value;
	}

	@Override
	public boolean equals(final Object o) {
		if (this == o) {
			return true;
		}
		if ((o == null) || (getClass() != o.getClass())) {
			return false;
		}
		ConfigSetting configSetting = (ConfigSetting) o;
		return Objects.equals(this.name, configSetting.name) &&
				Objects.equals(this.type, configSetting.type) &&
				Objects.equals(this.value, configSetting.value);
	}

	@Override
	public int hashCode() {
		return Objects.hash(name, type, value);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class ConfigSetting {\n");
		sb.append("    name: ").append(toIndentedString(name)).append("\n");
		sb.append("    type: ").append(toIndentedString(type)).append("\n");
		sb.append("    value: ").append(toIndentedString(value)).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(final Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}

}